+++ src/sql/parts/dashboard/contents/webviewContent.css
@@ -3,7 +3,8 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-dashboard-widget-tab {
+webview-content {
 	height: 100%;
-	width: 100%;
+	width : 100%;
+	display: block;
 }
