+++ src/sql/parts/dashboard/contents/widgetContent.component.html
