+++ src/sql/parts/dashboard/contents/widgetContent.css
@@ -3,7 +3,7 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-dashboard-widget-tab {
+widget-content {
 	height: 100%;
 	width: 100%;
 }
